requirements.txt
requests==2.32.3
pydantic==2.7.3
python-dotenv==1.0.1
pydantic-settings==2.6.1